Изменение архитектуры
